[Styleguide] rename Slider -> Carousel
